update: update process of delete extent
add package new type of opcode OpSplitMarkDelete to identify split delete extent,we don't delete whole extent while recive this opcode.
but it cann't cover extentKey already exist which might be delete on scenrio of snapshot deletion,
hence we judge according to scope of offset and size contrast with exist extent info,if scope cover the whole extent then delete the extent or else not.

Signed-off-by: leonrayang <chl696@sina.com>