update: fix extent write conflict on datanode
write on exist extent normally happened and create a new extent to write. but for snapshot scenarios extent be split
and the last extent to append write directly may occur conflict because the position may have the splited extent and effective,
thus we need make a additional jugement for GetEndForAppendW, if split happend then not reuse the extent.

Signed-off-by: leonrayang <changliang@oppo.com>